+++ krita/plugins/filters/colorsfilters/kis_perchannel_filter.cc
@@ -291,7 +291,7 @@ void KisPerChannelFilter::process(KisFilterConstProcessingInformation srcInfo,
                 // adjust, but since it's partially selected we also only partially adjust
                 adj->transform(iter.oldRawData(), iter.rawData(), 1);
                 const quint8 *pixels[2] = {iter.oldRawData(), iter.rawData()};
-                quint8 weights[2] = {MAX_SELECTED - selectedness, selectedness};
+                qint16 weights[2] = {MAX_SELECTED - selectedness, selectedness};
                 mixOp->mixColors(pixels, weights, 2, iter.rawData());
                 ++iter;
                 pixelsProcessed++;
