+++ krita/plugins/paintops/dynamicbrush/lib/kis_dynamic_coloring.cc
@@ -123,7 +123,7 @@ void KisPlainColoring::selectColor(double mix)
     colors[0] = m_cachedBackGroundColor->data();
     colors[1] = m_foreGroundColor.data();
     int weight = (int)(mix * 255);
-    const quint8 weights[2] = { 255 - weight, weight };
+    const qint16 weights[2] = { 255 - weight, weight };
     
     m_color->colorSpace()->mixColorsOp()->mixColors(colors, weights, 2, m_color->data());
